% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wildfire.R
\name{Wildfire}
\alias{Wildfire}
\alias{risk_score}
\title{Wild Fire Risk Assessment in Santa Barbara}
\usage{
risk_score(T, p, h) risk_level(risk_score)
}
\arguments{
\item{T}{Numeric, annual mean temperature in Celsius.}

\item{p}{Numeric, annual mean precipitation in mm.}

\item{h}{Numeric, annual mean humidity in percent (%).}
}
\value{
The \code{risk_score} function returns a numeric risk score based on the input parameters.

The \code{risk_level} function returns a character string indicating the risk level:
"no risk", "low risk", "medium risk", or "high risk"
}
\description{
This package includes functions to calculate the risk score for wild fires in
 Santa Barbara using environmental parameters such as temperature, precipitation,
 and humidity, and to classify this score into risk levels.

The main functions are:
\itemize{
  \item{\code{risk_score}}{: Calculates a numerical risk score based on input parameters}
  \item{\code{risk_level}}{: Classifies the calculated risk score into a risk level category}
}
}
\examples{
risk_score(25, 200, 45)

risk_level(320)

}
\author{
Author:
Livia Costa and Zoe Zhou
}
