% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire_data.R
\docType{data}
\name{wildfire_data}
\alias{wildfire_data}
\title{Wildfire Risk Assessment Data for Santa Barbara}
\format{
A data frame with 15 rows and 4 variables
\describe{
  \item{year}{Year of the data collection.}
  \item{t}{Mean temperature (°C).}
  \item{p}{Mean precipitation (mm).}
  \item{h}{Mean humidity (\%).}
}
}
\source{
Estimates based on real data available at different internet sources.
}
\usage{
wildfire_data
}
\description{
This dataset contains estimated data considering the most recent minimum and maximum temperatures,
humidity, and precipitation in Santa Barbara for 2023.
}
\keyword{datasets}
